ecommerce-pair5

+++ ecommerce-pair5/src/main/java/com/tobeto/ecommercepair5/services/dtos/responses/product/AddProductResponse.java
@@ -22,7 +22,7 @@ public class AddProductResponse {
 
     private int stock;
 
-    private int unitPrice;
+    private double unitPrice;
 
     private LocalDate addedDate;
